requirements.txt
pandas
numpy
dataclasses
unicodecsv
keras
tensorflow~=2.9.1
setuptools
matplotlib
tikzplotlib
sklearn
scikit-learn
# loreley~=0.1.0
gym
scipy
seaborn
asn1crypto
atari-py
cached-property
chardet
cryptography
importlib-metadata
keyring
keyrings.alt
pyglet
pyxdg
stable-baselines
wincertstore
Django~=3.2.12
djangorestframework
django-extensions
django-cors-headers
Werkzeug
pyOpenSSL
PyJWT
requests
